Tests/LibWeb/Text/expected/css/pseudo-elements-on-hover.txt

update: 16 create: 0
update: 78 create: 33
update: 16 create: 0
